core: move `VisibilityChecker` into its own module
