+++ pg_search/src/postgres/mod.rs
@@ -32,6 +32,7 @@ pub mod datetime;
 pub mod transaction;
 pub mod types;
 pub mod utils;
+pub mod visibility_checker;
 
 #[repr(u16)] // b/c that's what [`pg_sys::StrategyNumber`] is
 pub enum ScanStrategy {
